requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.13.1
python-dotenv==1.0.1

# Database
aiosqlite==0.22.1

# AI / LLM
langgraph==1.1.3

# HTTP / Web
httpx==0.28.1
playwright==1.58.0
beautifulsoup4==4.13.4

# Report Generation
fpdf2==2.8.7
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
jinja2==3.1.4
matplotlib==3.10.7

# Notifications
aiosmtplib==5.1.0

# Utilities
loguru==0.7.2
